+++ Bases de dados/Hashtags.csv
@@ -1,0 +1,259 @@
+,hashtags,pol_dummy
+0,#ChileViolatesHumanRights,1
+1,#ChileSeDesperto,1
+2,#EnChileTorturanViolanYMatan,1
+3,#ChileEnLlamas,1
+4,#EligeChile,1
+5,#ChileDespierto,1
+6,#SigoChilenosEnOctubre,1
+7,#ChilenosQueHablan,1
+8,#ChileViolatesHumanRigths,1
+9,#FrenteAmpliodeChile,1
+10,#VamosChile,1
+11,#EvidenciaCrisisChile,1
+12,#ChileEnResistencia,1
+13,#OctubreParaSeguirChilenos,1
+14,#NuevaConstitucionParaChile,1
+15,#ChileDeTodos,1
+16,#ChileNoEstaEnGuerra,1
+17,#ChileViolaDDDHH,1
+18,#ChileDespertóAS,1
+19,#ChileDespeto,1
+20,#ChileEnLaCIDH,1
+21,#MarchaMasGrandeDeChile3,1
+22,#EstoEstaPasandoEnChile,1
+23,#ChileWantsListeningSession,1
+24,#ChileDesperto,1
+25,#EnChileNodejanCiegos,1
+26,#ToquedeQuedaChile,1
+27,#AsesinosdeChile,1
+28,#ChileVamos,1
+29,#ChileMasJusto,1
+30,#ChileProtesta,1
+31,#ChileBajoDictadura,1
+32,#ChileEnDictaduraa,1
+33,#RecuperemosChile,1
+34,#ChileViolaDDHH,1
+35,#ChileViolaLosDDHH,1
+36,#ChileEnRebeldia,1
+37,#DespiertaChile,1
+38,#LaMarchaMasGrandeDeChile2,1
+39,#ChileEnCaos,1
+40,#EnChileNosDejanCiegos,1
+41,#ChileQuiereCambiosً,1
+42,#EstoPasaEnChileً,1
+43,#TerrorismoDeEstadoEnChile,1
+44,#Chilezuela,1
+45,#ChileSeAburrio,1
+46,#EstoPasoEnChile,1
+47,#quenospasaChile,1
+48,#ChileViolateHumanRights,1
+49,#ChileQuiereCambios,1
+50,#ChileSomosTodos,1
+51,#todosunidosPorChile,1
+52,#ChileViolatesHumanRigts,1
+53,#ChileNoQuiereMigajas,1
+54,#NoParemosChile,1
+55,#MarchaMásGrandeDeChile,1
+56,#DisolverCarabinerosdeChile,1
+57,#ChileSeCanso,1
+58,#LaMarchaMásGrandeChile,1
+59,#QueChileDecida,1
+60,#ChileProtests,1
+61,#VamosChilenos,1
+62,#CancelCOP25Chile,1
+63,#YoTambienSoyChileno,1
+64,#ONUHelpsChile,1
+65,#ChileYaDesperto,1
+66,#ChilenosPorLaPaz,1
+67,#ChileNOsevende,1
+68,#EstoPasaEnChile,1
+69,#pazParaChileً,1
+70,#YoCuidoChile,1
+71,#DespiertaChileDespierta,1
+72,#pazParaChile,1
+73,#TodosSomosChile,1
+74,#ChileViolatesHumansRights,1
+75,#ChileViolaLosDDDH,1
+76,#ElChileQueQueremos,1
+77,#ChileResiste,1
+78,#ChileEstaDespertando,1
+79,#ChileViolaLosDerechosHumanos,1
+80,#ChileEnHuelga,1
+81,#EnChileSeViolanLosDDHH,1
+82,#PrimaveradeChile,1
+83,#ChileDespierta,1
+84,#ChileNoSeRinde,1
+85,#ChileSinMiedo,1
+86,#LaMarchaMasGrandeDeChile,1
+87,#ChileEnDictadura,1
+88,#ChileSeCansoً,1
+89,#ChileEnCrisis,1
+90,#carabinerosdechile,1
+91,#chileenhuega,1
+92,#chilesedefiende,1
+93,#saboreachile,1
+94,#onuenchile,1
+95,#nadieparalizachile,1
+96,#chiledeperto,1
+97,#estoeschile,1
+98,#chileprotestas,1
+99,#chilenoscambiandoelclima,1
+100,#Protestaschile,1
+101,#chileviolateshumanrightsً,1
+102,#chileciudadanostvn,1
+103,#lagranmarchadechile,1
+104,#enchilehaycensura,1
+105,#huelgaenchile,1
+106,#Prayforchile,1
+107,#chilecambio,1
+108,#chilecracia,1
+109,#chileenpaz,1
+110,#chiledespertoensdl,1
+111,#fuerzachilenos,1
+112,#chileselevanta,1
+113,#ParoNacional21Novchile,1
+114,#chalecosamarilloschile,1
+115,#chilesomostodo,1
+116,#chileprotest,1
+117,#chilequierepaz,1
+118,#chilebajoataque,1
+119,#resistechile,1
+120,#estadodeemergenciachile,1
+121,#elchilequeviene,1
+122,#chilepreparado,1
+123,#despertochile,1
+124,#FueraTonkaCensuradora,1
+125,#FueraSalas,1
+126,#FueraPiñeraAhora,1
+127,#FueraMarioRozas,1
+128,#FueraRosas,1
+129,#FueraMañalich,1
+130,#FueraPiñera,1
+131,#ChadwickFuera,1
+132,#FueraCubillos,1
+133,#FueraLosMilitares,1
+134,#FueraRueda,1
+135,#FueraVallejo,1
+136,#FueraTontaTomicic,1
+137,#FueraRozas,1
+138,#FueraReginato,1
+139,#FueraJackson,1
+140,#FueraIsabelPla,1
+141,#FueraLosMilicos,1
+142,#CamilaFloresFuera,1
+143,#FueraCariola,1
+144,#FueraPiñeraDictador,1
+145,#FueraMarioRosas,1
+146,#FueraFrenteAmplio,1
+147,#FueraBoric,1
+148,#EvasionTodoslosdias,1
+149,#EvasionMasivaTodoElDiaً,1
+150,#EvasionTodoElDia,1
+151,#EvasionMasivaALas18,1
+152,#EvasionMasiva,1
+153,#EvasionMasivaTodoElDia,1
+154,#EvadiresRobar,1
+155,#EvadirEsLuchar,1
+156,#RenunciaGloriaHutt,1
+157,#RenunciaFontaine,1
+158,#RenunciaCamila,1
+159,#RenunciaSharp,1
+160,#RozasRenuncia,1
+161,#RenunciaPiñera,1
+162,#RenunciaMicco,1
+163,#RenunciaPiñeraCuliao,1
+164,#CariolaRenuncia,1
+165,#SiPiñeraRenuncia,1
+166,#RenunciaChadwick,1
+167,#ChadwickRenuncia,1
+168,#RenunciaChadwickAhora,1
+169,#RenunciaPiñeraQlo,1
+170,#RenunciaMilf,1
+171,#RenunciaPiñeraAhora,1
+172,#PiñeraRenuncia,1
+173,#RenunciaMarioRozas,1
+174,#RenunciaChanchoCuliao,1
+175,#RenunciaMañalich,1
+176,#RenunciaMartinez,1
+177,#RenunciaIsamit,1
+178,#RenunciaChadwck,1
+179,#PileraRenuncia,1
+180,#RenunciaPiñeraDictador,1
+181,#RenunciaCubillos,1
+182,#RenunciaPla,1
+183,#RenunciaRozas,1
+185,#PiñeraPayasoCuliao,1
+186,#ChadwickCuliao,1
+187,#PiñeraCuliao,1
+189,#PiñeraReculiao,1
+190,#pacosculiaos,1
+191,#PiñeraVerguenzaMundial,1
+192,#PiñeraTeQuedaPoco,1
+193,#ToqueDeQueda,1
+194,#ToqueDeQuedaYa,1
+196,#ToqueQueda,1
+199,#LaMarchaMasGrandeDeTodas,1
+200,#LaMarchaMasGrandeDeLaQuinta,1
+202,#JuntosSomosMásGrandes,1
+203,#LaMarchaMasGrandeDelMundo,1
+206,#LaMarchaMasGrandeDeTodas3,1
+207,#MarchaMasGrandeDeTodas,1
+209,#FacultadesACarabineros,1
+210,#CarabinerosViolaDDHH,1
+211,#CarabinerosAsesinos,1
+212,#TestDeDrogasParaCarabineros,1
+213,#CacerolazoNacional,1
+214,#Cacerolazo,1
+215,#CacerolazoPorGustavo,1
+216,#ToqueDeCacerolazo,1
+217,#CacerolazoPorLaSalud,1
+218,#CacerolazoLatinoamericano,1
+219,#cacerolazomasivo,1
+222,#PiñeraCriminal,1
+223,#AcusacionParaPiñeraYa,1
+224,#PeligroPiñeraMata,1
+226,#DestitucionDePiñera,1
+228,#DeletePiñera,1
+231,#PiñeraViolaDDH,1
+232,#PiñeraAsesino,1
+233,#MacriPiñeraLaMismaBilletera,1
+234,#PiñeraLies,1
+235,#PiñeraViolaDDHH,1
+236,#PiñeraEstoyContigo,1
+237,#PiñeraDictador,1
+238,#PiñeraViolatesHumanRights,1
+239,#LiderazgoDePiñera,1
+242,#dondeestaPiñera,1
+244,#PiñeraYaFue,1
+245,#PiñeraAcusado,1
+247,#PiñeraMiente,1
+248,#Piñera,1
+249,#AcusacionConstitucionalPiñera,1
+251,#DictadorPiñera,1
+254,#Piñeraesviolencia,1
+255,#DictaduraPiñera,1
+256,#PiñeraNoTeRobesLaMarcha,1
+257,#DestitucionParaPiñera,1
+258,#PiñeraVerguenzaNacional,1
+260,#CarcelParaPiñera,1
+261,#Piñera25Muertes,1
+262,#FuerzaPiñera,1
+263,#PiñeraEnCooperativa,1
+264,#DictaduradePiñera,1
+266,#somosmayoria,1
+267,#PacoLeaks,1
+268,#PrensaProstituida,1
+269,#SinIgualdadNoHayPaz,1
+270,#EstadoDeSitio,1
+271,#evasionmetro,1
+272,#evacionmasiva,1
+273,#17DeOctubre,1
+274,#PlazaItalia,1
+275,#EvacionMasivaALas1800,1
+276,#SeSabeQueEresPobreSi,1
+277,#SantiagoNoEsChile,1
+278,#DesobedienciaCivil,1
+279,#YaNoTengoMiedo,1
+280,#metrosantiago,1
+281,#metrodesantiago,1
